ai-finance-assistant/requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
pydantic>=2.6.0

langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.8
langgraph>=0.2.0

faiss-cpu>=1.8.0
tiktoken>=0.7.0

yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0

feedparser>=6.0.11
requests>=2.32.0

pytest>=8.2.0